consolidates partition behavior in IPartitioner, so creating a new partitioner should be only a matter of implementing that interface.  all the external switch statements on PartitionerType have been folded into that.
SSTable is now the only part of the code that cares about the distinction between a 'raw' key and a 'decorated' key.  variables in that class have been named clientKey or decoratedKey to show which is which.  others don't care either because they only deal with decorated keys (SequenceFile) or only with client keys (everyone else).  as part of this, I've merged some overloaded methods with substantially duplicated code to simplify auditing these changes.

git-svn-id: http://svn.apache.org/repos/asf/incubator/cassandra/trunk@758997 13f79535-47bb-0310-9956-ffa450edef68